requirements.txt
streamlit
langchain
chromadb
faiss-cpu
openai
sentence-transformers
ollama
python-dotenv
PyMuPDF>=1.22.0
pdfplumber
python-pptx
python-docx
